lab management range decimal change

+++ GSC.Data/Dto/LabManagement/LabManagementVariableMappingDto.cs
@@ -21,10 +21,10 @@ namespace GSC.Data.Dto.LabManagement
     {
         public int ProjectDesignVariableId { get; set; }
         public string TargetVariable { get; set; }
-        public int? MaleLowRange { get; set; }
-        public int? MaleHighRange { get; set; }
-        public int? FemaleLowRange { get; set; }
-        public int? FemaleHighRange { get; set; }
+        public decimal? MaleLowRange { get; set; }
+        public decimal? MaleHighRange { get; set; }
+        public decimal? FemaleLowRange { get; set; }
+        public decimal? FemaleHighRange { get; set; }
         public string Unit { get; set; }
     }
